// generated/src/app/api-models/cap-rate-detail.model.ts
/* Copyright (c) 2020 . All Rights Reserved. */

export class CapRateDetail {
  seqCapRateHdr: number;
  seqAgeBandDtl: number;
  capRate: number;
  securityCode: string;
  insertDatetime: Date;
  insertUser: string;
  insertProcess: string;
  updateDatetime: Date;
  updateUser: string;
  updateProcess: string;
}